refactor: Modularize application startup configuration
- Extracted key application setup steps into dedicated configuration files:
  - Database configuration (`configureDatabase`)
  - CORS policy (`configureCors`)
  - GraphQL setup (`configureGraphQL`)
  - Route definitions (`configureRouting`)
  - MCP server configuration (`configureMcp`)
  - Compression and caching (`configureCompression`, `configureCaching`)
  - Call logging (`configureCallLogging`)
  - Status pages (`configureStatusPages`)
- Simplified `Application.module()` by delegating to extracted functions.
